backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
supabase==2.3.4
sqlalchemy==2.0.23
asyncpg==0.29.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and validation
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client and utilities
httpx==0.25.2
python-dateutil==2.8.2

# Additional utilities (built-in modules, no installation needed)
# secrets - For secure token generation (built-in)
# logging - For structured logging (built-in)  
# pathlib - For file path handling (built-in)

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Production dependencies (install separately if needed)
# gunicorn==21.2.0  # Production WSGI server
# psycopg2-binary==2.9.9  # PostgreSQL adapter (requires PostgreSQL dev tools)